refactor: fold certs, error_pages, and render into nginx __init__
NginxConfigRenderer, the cert-path helpers, and the error-page renderer
each had a single caller (NginxManager, sometimes both). Collapsing them
into one module removes the split without changing any behavior, and
lets NginxManager and _ensure_modsecurity_module/_set_worker_user/
install_default_server share one _stage_and_copy helper for the
stage-then-sudo-copy pattern instead of repeating it three times.
